# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/san-pham": {
      "post": {
        "description": "Tạo một sản phẩm mới với thông tin đầy đủ bao gồm hình ảnh",
        "operationId": "SanPhamController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSanPhamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sản phẩm được tạo thành công"
          },
          "400": {
            "description": "Dữ liệu đầu vào không hợp lệ"
          },
          "404": {
            "description": "Nhân viên hoặc danh mục không tồn tại"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo sản phẩm mới",
        "tags": [
          "Sản phẩm"
        ]
      },
      "get": {
        "description": "Trả về danh sách tất cả sản phẩm với thông tin chi tiết",
        "operationId": "SanPhamController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lấy danh sách sản phẩm thành công"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy tất cả sản phẩm",
        "tags": [
          "Sản phẩm"
        ]
      }
    },
    "/san-pham/page/{stt}": {
      "get": {
        "description": "Phân trang danh sách sản phẩm, mỗi trang 20 sản phẩm",
        "operationId": "SanPhamController_findByPage",
        "parameters": [
          {
            "name": "stt",
            "required": true,
            "in": "path",
            "description": "Số trang (bắt đầu từ 1)",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy sản phẩm theo trang thành công"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy sản phẩm theo trang",
        "tags": [
          "Sản phẩm"
        ]
      }
    },
    "/san-pham/search": {
      "get": {
        "description": "Tìm kiếm sản phẩm dựa trên tên sản phẩm",
        "operationId": "SanPhamController_findByName",
        "parameters": [
          {
            "name": "ten",
            "required": true,
            "in": "query",
            "description": "Tên sản phẩm cần tìm",
            "schema": {
              "example": "iPhone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tìm kiếm thành công"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tìm kiếm sản phẩm theo tên",
        "tags": [
          "Sản phẩm"
        ]
      }
    },
    "/san-pham/danh-muc/{danhMucId}": {
      "get": {
        "description": "Trả về danh sách sản phẩm thuộc một danh mục cụ thể",
        "operationId": "SanPhamController_findByDanhMuc",
        "parameters": [
          {
            "name": "danhMucId",
            "required": true,
            "in": "path",
            "description": "ID danh mục",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy sản phẩm theo danh mục thành công"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy sản phẩm theo danh mục",
        "tags": [
          "Sản phẩm"
        ]
      }
    },
    "/san-pham/{id}": {
      "get": {
        "description": "Trả về thông tin chi tiết của một sản phẩm bao gồm hình ảnh và đánh giá",
        "operationId": "SanPhamController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID sản phẩm",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin sản phẩm thành công"
          },
          "404": {
            "description": "Không tìm thấy sản phẩm"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin chi tiết sản phẩm",
        "tags": [
          "Sản phẩm"
        ]
      },
      "delete": {
        "description": "Xóa sản phẩm khỏi hệ thống",
        "operationId": "SanPhamController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID sản phẩm cần xóa",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa sản phẩm thành công"
          },
          "404": {
            "description": "Không tìm thấy sản phẩm"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa sản phẩm",
        "tags": [
          "Sản phẩm"
        ]
      }
    },
...
    "/taikhoan/{id}": {
      "get": {
        "operationId": "TaiKhoanController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TaiKhoan"
        ]
      },
      "patch": {
        "operationId": "TaiKhoanController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaiKhoanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TaiKhoan"
        ]
      },
      "delete": {
        "operationId": "TaiKhoanController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TaiKhoan"
        ]
      }
    },
...
    "/danh-muc/{tendanhmuc}": {
      "get": {
        "operationId": "DanhMucController_findOne",
        "parameters": [
          {
            "name": "tendanhmuc",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DanhMuc"
        ]
      }
    },
    "/danh-muc/{id}": {
      "patch": {
        "operationId": "DanhMucController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDanhMucDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DanhMuc"
        ]
      },
      "delete": {
        "operationId": "DanhMucController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DanhMuc"
        ]
      }
    },
...
    "/danh-gia/san-pham/{sanPhamId}": {
      "get": {
        "operationId": "DanhGiaController_getBySanPham",
        "parameters": [
          {
            "name": "sanPhamId",
            "required": true,
            "in": "path",
            "description": "ID sản phẩm",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách thành công"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy đánh giá theo ID sản phẩm",
        "tags": [
          "Đánh giá"
        ]
      }
    },
    "/danh-gia/khach-hang/{khachHangId}": {
      "get": {
        "operationId": "DanhGiaController_getByKhachHang",
        "parameters": [
          {
            "name": "khachHangId",
            "required": true,
            "in": "path",
            "description": "ID khách hàng",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách thành công"
          },
          "403": {
            "description": "Không có quyền truy cập"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy đánh giá theo ID khách hàng",
        "tags": [
          "Đánh giá"
        ]
      }
    },
...
    "/thong-tin-lien-he": {
      "post": {
        "operationId": "ThongTinLienHeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThongTinLienHeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ThongTinLienHe"
        ]
      },
      "get": {
        "operationId": "ThongTinLienHeController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ThongTinLienHe"
        ]
      }
    },
    "/thong-tin-lien-he/{id}": {
      "get": {
        "operationId": "ThongTinLienHeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ThongTinLienHe"
        ]
      },
      "patch": {
        "operationId": "ThongTinLienHeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThongTinLienHeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ThongTinLienHe"
        ]
      },
      "delete": {
        "operationId": "ThongTinLienHeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ThongTinLienHe"
        ]
      }
    },
    "/khach-hang": {
      "post": {
        "operationId": "KhachHangController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKhachHangDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đăng ký khách hàng",
        "tags": [
          "KhachHang"
        ]
      },
      "get": {
        "operationId": "KhachHangController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KhachHang"
        ]
      }
    },
    "/khach-hang/{id}": {
      "get": {
        "operationId": "KhachHangController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KhachHang"
        ]
      },
      "delete": {
        "operationId": "KhachHangController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KhachHang"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Quản lý Siêu Thị",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateSanPhamDto": {
        "type": "object",
        "properties": {
          "ten": {
            "type": "string"
          },
          "soLuongHienCon": {
            "type": "number"
          },
          "hanSuDung": {
            "format": "date-time",
            "type": "string"
          },
          "ngaySanXuat": {
            "format": "date-time",
            "type": "string"
          },
          "moTa": {
            "type": "string"
          },
          "gia": {
            "type": "number"
          },
          "danhMucId": {
            "type": "number",
            "description": "ID của danh mục sản phẩm"
          },
          "hinhAnhs": {
            "description": "Danh sách hình ảnh sản phẩm",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ten",
          "soLuongHienCon",
          "hanSuDung",
          "ngaySanXuat",
          "moTa",
          "gia",
          "danhMucId"
        ]
      },
...
      "UpdateTaiKhoanDto": {
        "type": "object",
        "properties": {
          "tenDangNhap": {
            "type": "string"
          },
          "matKhau": {
            "type": "string"
          }
        }
      },
...
      "UpdateDanhMucDto": {
        "type": "object",
        "properties": {
          "tenDanhMuc": {
            "type": "string",
            "description": "Tên danh mục"
          },
          "moTa": {
            "type": "string",
            "description": "Mô tả danh mục"
          }
        }
      },
...
      "CreateThongTinLienHeDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateThongTinLienHeDto": {
        "type": "object",
        "properties": {}
      },
      "CreateKhachHangDto": {
        "type": "object",
        "properties": {
          "tenKhachHang": {
            "type": "string"
          },
          "sdt": {
            "type": "string"
          },
          "tenDangNhap": {
            "type": "string"
          },
          "matKhau": {
            "type": "string"
          }
        },
        "required": [
          "tenKhachHang",
          "sdt",
          "tenDangNhap",
          "matKhau"
        ]
      },
...
    }
  }
}